+++ 09.wc_object/main.rb
@@ -2,9 +2,9 @@
 
 require 'optparse'
 require_relative './lib/calc'
-require_relative './lib/wc_input'
+require_relative './lib/input_calculator'
 require_relative './lib/wc_command'
-require_relative './lib/wc_files'
+require_relative './lib/files_calculator'
 
 options = ARGV.getopts('l')
 wc_command = WcCommand.new
